+++ bot/__init__.py
@@ -1,3 +1,3 @@
 import telebot
+import spamwatch
 from .load_environ import Environment
-from .check_bot_permission import Bot_Permissions