+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [unreleased]
-- Initial release
+- Added bot required permission checker
+- Added [SpamWatch API](https://github.com/SpamWatch) to filter user
 
 ---
 [unreleased]: https://github.com/BiltuDas1/vigilantbot-telegram